chore(libvmaf): CodeQL C bulk sweep — 60/78 deferred alerts fixed
Resolves the 60 mechanically-fixable alerts deferred from #538 across three
CodeQL categories on upstream-mirror C files:

* cpp/integer-multiplication-cast-to-long (44/44):
  prefixed widening cast on the LHS operand before multiplication —
  (size_t), (uint64_t), (double), (long), (ptrdiff_t) — never wrapped
  around the whole expression. Touched y4m_input.c (13), cambi.c (5),
  integer_psnr.c (4), iqa/convolve.c (deferred — bit-exactness gate),
  vmaf_bench.c (4), psnr_hvs.c (2), ansnr_avx2.c (2), ansnr_avx512.c (2),
  ciede.c (2), pdjson.c (1), yuv_input.c (1), adm_tools.h (1),
  integer_adm.c (1), adm_avx2.c (1), adm_avx512.c (1).
* cpp/commented-out-code (19/19): deleted dead/disabled code blocks
  (TODO-with-corpse-code, debug-print scaffolding, AVX-512 VP-loop dead
  variant, libsvm dead alternate). Per CLAUDE no-corpse-comments rule.
* cpp/large-parameter (1/15): converted cambi_score(... CambiBuffers
  buffers, ...) → const CambiBuffers *. The 14 remaining alerts on the
  VifBuffer SIMD ABI (integer_vif.{c,h}, x86/vif_avx2.c, x86/vif_avx512.c)
  are deferred to a coordinated multi-backend follow-up — the calling
  convention is shared with arm64/vif_neon.c (not in this sweep) and
  flipping it requires lockstep updates across NEON + AVX2 + AVX-512 +
  scalar paths.

Touched-file lint-clean per CLAUDE §12 r12. One unrelated IWYU warning
absorbed: dropped unused config.h include in libvmaf/test/test_cpu.c
after deleting the AVX2 mask-test corpse that gated it. Build + 50 C
unit tests pass on CPU configuration. Netflix golden gate runs identical
9 pre-existing failures as master (PSNR/NIQE Python harness, unrelated).

Coordinates with parallel sibling sweep on cpp/declaration-hides-variable
(64 alerts, branch agent-a5fe363b26a4d7592) — touches many of the same
files; whichever lands first, the other rebases.

+++ libvmaf/src/feature/ciede.c
@@ -282,7 +282,8 @@ static float ciede2000(LABColor color_1, LABColor color_2, KSubArgs ksub)
     const float h_prime_1 = get_h_prime(color_1.b, a_prime_1);
     const float h_prime_2 = get_h_prime(color_2.b, a_prime_2);
     const float delta_h_prime = get_delta_h_prime(c1, c2, h_prime_1, h_prime_2);
-    const float delta_upcase_h_prime = 2.0 * sqrt(c_prime_1 * c_prime_2) * sin(delta_h_prime / 2.0);
+    const float delta_upcase_h_prime =
+        2.0 * sqrt((double)c_prime_1 * c_prime_2) * sin(delta_h_prime / 2.0);
     const float upcase_h_bar_prime = get_upcase_h_bar_prime(h_prime_1, h_prime_2);
     const float upcase_t = get_upcase_t(upcase_h_bar_prime);
     const float s_sub_upcase_h = 1.0 + 0.015 * c_bar_prime * upcase_t;
@@ -291,7 +292,7 @@ static float ciede2000(LABColor color_1, LABColor color_2, KSubArgs ksub)
     const float chroma = delta_c_prime / (ksub.c * s_sub_c);
     const float hue = delta_upcase_h_prime / (ksub.h * s_sub_upcase_h);
 
-    return sqrt(pow(lightness, 2) + pow(chroma, 2) + pow(hue, 2) + r_sub_t * chroma * hue);
+    return sqrt(pow(lightness, 2) + pow(chroma, 2) + pow(hue, 2) + (double)r_sub_t * chroma * hue);
 }
 
 static double pow_2_4(double x)
